markdown_ai_cli/requirements.txt
typer
langchain
langchain-openai
python-dotenv